navbar, moved login/signup
