+++ README.md
@@ -5,7 +5,6 @@ A project by:
 Micaela O'Herron
 Maria Sanchez
 
-
 _Good things come in pairs_
 
 Looking to mix up a backend with express/sequelize and a frontend with react/redux? That's `boilermaker`!
