+++ client/components/about.js
@@ -1,10 +1,10 @@
 import React from 'react'
 
-const About = (props) => {
-    return (
-        //put jsx here
-        <div></div>
-    )
+const About = props => {
+  return (
+    //put jsx here
+    <h1>About Grace's Cakes</h1>
+  )
 }
 
 export default About
