+++ client/components/adminHome.js
@@ -1,12 +1,7 @@
 import React from 'react'
 
-
-const AdminHome = (props) => {
-  
-  return (
-    <div></div>
-  )
+const AdminHome = props => {
+  return <div />
 }
-
 
 export default AdminHome
