+++ client/components/home.js
@@ -1,10 +1,9 @@
 import React from 'react'
 
-const Home = (props) => {
-    return (
-        //put jsx here
-        <div></div>
-    )
+const Home = props => {
+  return (
+    <h1>Welcome to our bakery!</h1>
+  )
 }
 
 export default Home
