+++ client/components/myAccount.js
@@ -1,25 +1,51 @@
 import React from 'react'
+import PropTypes from 'prop-types'
+import {connect} from 'react-redux'
+import {logout} from '../store'
+import {Link} from 'react-router-dom'
 
-const MyAccount = (props) => {
-    return (
-        isLoggedIn ? (
-            <div>
-              {/* The navbar will show these links after you log in */}
-              <Link to="/home">Home</Link>
-              <a href="#" onClick={handleClick}>
-                Logout
-              </a>
-            </div>
-          ) : (
-            <div>
-              {/* The navbar will show these links before you log in */}
-              <Link to="/login">Login</Link>
-              <Link to="/signup">Sign Up</Link>
-            </div>
-        )
-    )
+const MyAccount = ({handleClick, isLoggedIn}) => {
+  return (
+    <div>
+      {isLoggedIn ? (
+        <div>
+          {/* The navbar will show these links after you log in */}
+          <Link to="/home">Home</Link>
+          <a href="#" onClick={handleClick}>
+            Logout
+          </a>
+        </div>
+      ) : (
+        <div>
+          {/* The navbar will show these links before you log in */}
+          <Link to="/login">Login</Link>
+          <Link to="/signup">Sign Up</Link>
+        </div>
+      )}
+    </div>
+  )
 }
 
-//TODO get isLoggedIn
+const mapState = state => {
+  return {
+    isLoggedIn: !!state.user.id
+  }
+}
 
-export default MyAccount
+const mapDispatch = dispatch => {
+  return {
+    handleClick() {
+      dispatch(logout())
+    }
+  }
+}
+
+/**
+ * PROP TYPES
+ */
+MyAccount.propTypes = {
+  handleClick: PropTypes.func.isRequired,
+  isLoggedIn: PropTypes.bool.isRequired
+}
+
+export default connect(mapState, mapDispatch)(MyAccount)
