+++ client/components/shop.js
@@ -1,13 +1,12 @@
 import React, {Component} from 'react'
 
 class Shop extends Component {
-
-    render() {
-        return (
-            //list all products here
-            <div></div>
-        )
-    }
+  render() {
+    return (
+      //list all products here
+      <div />
+    )
+  }
 }
 
 export default Shop
